SimCLR-2/requirements.txt
numpy
pandas
scikit-image
torch
torchvision
tensorboard
onnx
onnxruntime
tqdm
PyYAML
opencv-python